+++ test/pow/phoenix/controllers/controller_test.exs
@@ -22,5 +22,47 @@ defmodule Pow.Phoenix.ControllerTest do
       assert html = html_response(conn, 200)
       assert html =~ ":web_module new session"
     end
+
+    test "with halt {:halt, conn} from before_process", %{conn: conn} do
+      defmodule HaltsBeforeProcess do
+        use Pow.Extension.Phoenix.ControllerCallbacks.Base
+
+        def before_process(SessionController, :new, conn, _config) do
+          {:halt, Conn.halt(conn)}
+        end
+
+        def before_respond(SessionController, :new, _, _config) do
+          raise "Should not be called"
+        end
+      end
+
+      conn =
+        conn
+        |> Conn.put_private(:pow_config, user: User, controller_callbacks: HaltsBeforeProcess)
+        |> Conn.put_private(:phoenix_action, :new)
+
+      conn = Controller.action(SessionController, conn, %{})
+
+      assert conn.halted
+    end
+
+    test "{:halt, conn} from before_respond", %{conn: conn} do
+      defmodule HaltsBeforeRespond do
+        use Pow.Extension.Phoenix.ControllerCallbacks.Base
+
+        def before_respond(SessionController, :new, {:ok, _changeset, conn}, _config) do
+          {:halt, Conn.halt(conn)}
+        end
+      end
+
+      conn =
+        conn
+        |> Conn.put_private(:pow_config, user: User, controller_callbacks: HaltsBeforeRespond)
+        |> Conn.put_private(:phoenix_action, :new)
+
+      conn = Controller.action(SessionController, conn, %{})
+
+      assert conn.halted
+    end
   end
 end